all classes changed the map from integer to double in all classes with coffee. html coffeType + css

+++ CoffiiMe/src/main/java/app/coffeeTypes/americano/Americano.java
@@ -8,9 +8,9 @@ public class Americano implements Coffee {
 
     private Color color;
     private int volume;
-    private Map<String, Integer> ingredients;
+    private Map<String, Double> ingredients;
 
-    public Americano(Color color, int volume, Map<String, Integer> ingredients) {
+    public Americano(Color color, int volume, Map<String, Double> ingredients) {
         this.color = color;
         this.volume = volume;
         this.ingredients = ingredients;
@@ -27,7 +27,7 @@ public class Americano implements Coffee {
     }
 
     @Override
-    public Map<String, Integer> getIngredients() {
+    public Map<String, Double> getIngredients() {
         return ingredients;
     }
 }
